refactor: improve first visit modal
