+++ bridge-ui/src/app/buy/page.tsx
@@ -2,15 +2,13 @@
 
 import OnRamperWidget from "@/components/onramper";
 import styles from "./page.module.scss";
-import useFirstVisitModal from "@/hooks/useFirstVisitModal";
+import FirstVisitModal from "@/components/modal/first-time-visit";
 
 export default function Page() {
-  const modal = useFirstVisitModal({ type: "buy" });
-
   return (
     <section className={styles["content-wrapper"]}>
       <OnRamperWidget />
-      {modal}
+      <FirstVisitModal type="buy" />
     </section>
   );
 }
