+++ bridge-ui/src/components/modal/first-time-visit/index.tsx
@@ -1,3 +1,4 @@
+import { useCallback, useEffect, useMemo, useState } from "react";
 import Modal from "@/components/modal";
 import styles from "./first-time-visit.module.scss";
 import Button from "@/components/ui/button";
@@ -5,7 +6,9 @@ import Image from "next/image";
 import LineaIcon from "@/assets/logos/linea.svg";
 import CloseIcon from "@/assets/icons/close.svg";
 
-export type FirstTimeModalDataType = {
+type ModalType = "all-bridges" | "native-bridge" | "buy";
+
+type FirstTimeModalDataType = {
   title: string;
   description: string;
   steps: string[];
@@ -18,36 +21,103 @@ export type FirstTimeModalDataType = {
   };
 };
 
-type Props = {
+const modalData: Record<ModalType, FirstTimeModalDataType> = {
+  "all-bridges": {
+    title: "Welcome to the Linea Bridge!",
+    description: "Move your funds to Linea through the fastest route, at the lowest cost, and with no extra fees!",
+    steps: [
+      "Select your source chain & token",
+      "Choose Linea as your destination",
+      "Enter the amount & get the best rate",
+      "Connect your wallet & bridge",
+      "Your funds land on Linea in seconds",
+    ],
+    btnText: "Start bridging now",
+    extraText: "Ready to bridge?",
+    image: {
+      src: "/images/illustration/bridge-first-time-modal-illustration.svg",
+      width: 128,
+      height: 179,
+    },
+  },
+  "native-bridge": {
+    title: "Welcome to the Native Bridge!",
+    description:
+      "Ethereum to Linea using Linea’s official bridge. No third parties, no extra fees—just a direct way to move your assets.",
+    steps: [
+      "Select the token and amount you want to bridge from Ethereum to Linea.",
+      "Connect your wallet & approve",
+      "Confirm and wait - your funds land on Linea in about 20 minutes",
+    ],
+    btnText: "Start bridging now",
+    extraText: "Ready to bridge?",
+    image: {
+      src: "/images/illustration/bridge-first-time-modal-illustration.svg",
+      width: 128,
+      height: 179,
+    },
+  },
+  buy: {
+    title: "Fund Your Linea Wallet",
+    description:
+      "Buy tokens instantly at the best rates and with no extra fees. We compare multiple providers to find you the best rates and fastest transactions.",
+    steps: [
+      "Pick a token & amount",
+      "Select a payment method (card, bank, etc.) and follow the instruction",
+      "Connect your wallet",
+      "Confirm & receive tokens in seconds",
+    ],
+    btnText: "Buy tokens now",
+    image: {
+      src: "/images/illustration/buy-first-time-modal-illustration.svg",
+      width: 157,
+      height: 167,
+    },
+  },
+};
+
+export default function FirstVisitModal({ type }: { type: ModalType }) {
+  const [showModal, setShowModal] = useState(false);
+  const [shouldRenderModal, setShouldRenderModal] = useState(false);
+
+  const data = useMemo(() => modalData[type], [type]);
+
+  useEffect(() => {
+    if (typeof window !== "undefined" && localStorage.getItem(`hasVisited-${type}`) !== "true") {
+      setShowModal(true);
+      setShouldRenderModal(true);
+    }
+  }, [type]);
+
+  const handleClose = useCallback(() => {
+    setShowModal(false);
+    setTimeout(() => {
+      setShouldRenderModal(false);
+    }, 300);
+    if (typeof window !== "undefined") {
+      localStorage.setItem(`hasVisited-${type}`, "true");
+    }
+  }, [type]);
+
+  return shouldRenderModal ? <BaseModal isModalOpen={showModal} onCloseModal={handleClose} data={data} /> : null;
+}
+
+type BaseModalProps = {
   isModalOpen: boolean;
   onCloseModal: () => void;
   data: FirstTimeModalDataType;
 };
 
-export default function FirstTimeVisit({ isModalOpen, onCloseModal, data }: Props) {
+function BaseModal({ isModalOpen, onCloseModal, data }: BaseModalProps) {
   const { title, description, steps, btnText, extraText, image } = data;
-  const modalHeader = (
-    <div className={styles["header-wrapper"]}>
-      <Image
-        className={styles.illustration}
-        src={image.src}
-        width={image.width}
-        height={image.height}
-        role="presentation"
-        alt="modal image"
-      />
-      <div className={styles["close-icon"]} onClick={onCloseModal} role="button">
-        <CloseIcon />
-      </div>
-      <div className={styles["header-content"]}>
-        <LineaIcon className={styles.logo} />
-        <h3 className={styles.title}>{title}</h3>
-      </div>
-    </div>
-  );
 
   return (
-    <Modal title={title} isOpen={isModalOpen} onClose={onCloseModal} modalHeader={modalHeader}>
+    <Modal
+      title={title}
+      isOpen={isModalOpen}
+      onClose={onCloseModal}
+      modalHeader={<ModalHeader image={image} title={title} onCloseModal={onCloseModal} />}
+    >
       <div className={styles["modal-inner"]}>
         <p className={styles.description}>{description}</p>
         <p className={styles.how}>How it works:</p>
@@ -65,5 +135,32 @@ export default function FirstTimeVisit({ isModalOpen, onCloseModal, data }: Prop
         </Button>
       </div>
     </Modal>
+  );
+}
+
+type ModalHeaderProps = {
+  onCloseModal: () => void;
+  title: FirstTimeModalDataType["title"];
+  image: FirstTimeModalDataType["image"];
+};
+
+function ModalHeader({ image, title, onCloseModal }: ModalHeaderProps) {
+  return (
+    <div className={styles["header-wrapper"]}>
+      <Image
+        className={styles.illustration}
+        src={image.src}
+        width={image.width}
+        height={image.height}
+        alt="modal image"
+      />
+      <div className={styles["close-icon"]} onClick={onCloseModal} role="button" aria-label="Close modal">
+        <CloseIcon />
+      </div>
+      <div className={styles["header-content"]}>
+        <LineaIcon className={styles.logo} />
+        <h3 className={styles.title}>{title}</h3>
+      </div>
+    </div>
   );
 }
